리액트 라우터로 App.js 는 path에 따라 컴포넌트 그리도록 변경
