+++ movie/src/App.js
@@ -1,37 +1,21 @@
-import {useEffect, useState} from "react";
-import Movie from "./components/Movie";
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  Link
+} from 'react-router-dom'
+import Home from "./routes/Home";
+import Detail from "./routes/Detail";
 
 function App() {
-  const [loading, setLoading] = useState(true)
-  const [movies, setMovies] = useState([])
-  const getMovies = async () => {
-    const response = await fetch(
-      `https://yts.mx/api/v2/list_movies.json?minimum_rating=9&sort_by=year`
-    )
-    const json = await response.json()
-    setMovies(json.data.movies)
-    setLoading(false)
-  }
-  useEffect(() => {
-    getMovies()
-  }, [])
-
-  console.log(movies)
-
-  return <div>
-    {loading ? <h1>Loading...</h1>
-      : <div>
-        {movies.map(movie =>
-          <Movie
-            key={movie.id} // map 등 순회시 key 값은 필수
-            coverImg={movie.medium_cover_image}
-            title={movie.title}
-            summary={movie.summary}
-            genres={movie.genres}/>
-        )}
-      </div>
-    }
-  </div>
+  return (
+    <Router>
+      <Routes>
+        <Route path='/' element={<Home />} />
+        <Route path='/movie' element={<Detail />} />
+      </Routes>
+    </Router>
+  )
 }
 
 export default App
