+++ movie/src/components/Movie.js
@@ -1,10 +1,13 @@
 import PropTypes from "prop-types";
+import { Link } from 'react-router-dom'
 
 function Movie({ coverImg, title, summary, genres }) {
   return (
     <div>
       <img src={coverImg} alt={title} />
-      <h2>{title}</h2>
+      <h2>
+        <Link to="/movie">{title}</Link>
+      </h2>
       <p>{summary}</p>
       <ul>
         <li>
