+++ movie/src/routes/Detail.js
@@ -1,0 +1,5 @@
+function Detail() {
+  return <h1>Detail</h1>
+}
+
+export default Detail
